+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Bug Fixes
 
-* labels not being created as expected ([8030fb0](https://github.com/qomodo-labs/action-trissue/commit/8030fb06bb780abf71b3d1077ccd371248a7b598))
+- labels not being created as expected
+  ([8030fb0](https://github.com/qomodo-labs/action-trissue/commit/8030fb06bb780abf71b3d1077ccd371248a7b598))
 
 ## 1.0.0 (2025-03-25)